+++ R/scale_color_metro.R
@@ -1,9 +1,9 @@
 #' Metro color palette
 #'
-#' The palette based on Metro [Metro
-#' colors](https://materialui.co/metrocolors).
-#' Use `scale_color_metro_d` for *discrete* categories and
-#' `scale_color_metro_c` for a *continuous* scale.
+#' The palette based on Metro [Metro colors](https://materialui.co/metrocolors).
+#' Use `scale_color_metro_d` for *discrete* categories and `scale_color_metro_c`
+#' for a *continuous* scale, or use the `discrete` argument in
+#' `scale_color_metro()`.
 #'
 #' @inheritParams palette_metro
 #' @inheritParams scale_color_flat
@@ -15,23 +15,31 @@
 #' ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
 #'   geom_boxplot() +
 #'   theme_modern() +
-#'   scale_fill_metro_d()
+#'   scale_fill_metro()
 #'
 #' ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
 #'   geom_violin() +
 #'   theme_modern() +
-#'   scale_fill_metro_d(palette = "ice")
+#'   scale_fill_metro(palette = "ice")
 #'
 #' ggplot(iris, aes(x = Petal.Length, y = Petal.Width, color = Sepal.Length)) +
 #'   geom_point() +
 #'   theme_modern() +
-#'   scale_color_metro_c(palette = "rainbow")
+#'   scale_color_metro(discrete = FALSE)
 #' @export
-scale_color_metro <- function(palette = "complement",
+scale_color_metro <- function(palette = NULL,
                               discrete = TRUE,
                               reverse = FALSE,
                               aesthetics = "color",
                               ...) {
+  if (is.null(palette)) {
+    if (discrete) {
+      palette <- "contrast"
+    } else {
+      palette <- "gradient"
+    }
+  }
+
   pal <- palette_metro(palette = palette, reverse = reverse)
 
   if (discrete) {
@@ -47,7 +55,7 @@ scale_color_metro <- function(palette = "complement",
 
 #' @rdname scale_color_metro
 #' @export
-scale_color_metro_d <- function(palette = "complement",
+scale_color_metro_d <- function(palette = NULL,
                                 discrete = TRUE,
                                 reverse = FALSE,
                                 aesthetics = "color",
@@ -63,7 +71,7 @@ scale_color_metro_d <- function(palette = "complement",
 
 #' @rdname scale_color_metro
 #' @export
-scale_color_metro_c <- function(palette = "complement",
+scale_color_metro_c <- function(palette = NULL,
                                 discrete = FALSE,
                                 reverse = FALSE,
                                 aesthetics = "color",
@@ -95,11 +103,19 @@ scale_colour_metro_d <- scale_color_metro_d
 
 #' @rdname scale_color_metro
 #' @export
-scale_fill_metro <- function(palette = "complement",
+scale_fill_metro <- function(palette = NULL,
                              discrete = TRUE,
                              reverse = FALSE,
                              aesthetics = "fill",
                              ...) {
+  if (is.null(palette)) {
+    if (discrete) {
+      palette <- "contrast"
+    } else {
+      palette <- "gradient"
+    }
+  }
+
   pal <- palette_metro(palette = palette, reverse = reverse)
 
   if (discrete) {
@@ -112,7 +128,7 @@ scale_fill_metro <- function(palette = "complement",
 
 #' @rdname scale_color_metro
 #' @export
-scale_fill_metro_d <- function(palette = "complement",
+scale_fill_metro_d <- function(palette = NULL,
                                discrete = TRUE,
                                reverse = FALSE,
                                aesthetics = "fill",
@@ -128,7 +144,7 @@ scale_fill_metro_d <- function(palette = "complement",
 
 #' @rdname scale_color_metro
 #' @export
-scale_fill_metro_c <- function(palette = "complement",
+scale_fill_metro_c <- function(palette = NULL,
                                discrete = FALSE,
                                reverse = FALSE,
                                aesthetics = "fill",
@@ -148,18 +164,18 @@ scale_fill_metro_c <- function(palette = "complement",
 
 # The palette based on metro design colors: https://www.materialui.co/metrocolors
 metro_colors_list <- c(
-  `red` = "#e51400",
+  red = "#e51400",
   `dark red` = "#a20025",
-  `purple` = "#aa00ff",
+  purple = "#aa00ff",
   `deep purple` = "#76608a",
-  `blue` = "#0050ef",
+  blue = "#0050ef",
   `light blue` = "#1ba1e2",
-  `teal` = "#00aba9",
-  `green` = "#008a00",
+  teal = "#00aba9",
+  green = "#008a00",
   `light green` = "#60a917",
-  `yellow` = "#e3c800",
-  `amber` = "#f0a30a",
-  `orange` = "#fa6800",
+  yellow = "#e3c800",
+  amber = "#f0a30a",
+  orange = "#fa6800",
   `deep orange` = "#a0522d",
   `blue grey` = "#647687"
 )
@@ -191,9 +207,10 @@ metro_colors <- function(...) {
 
 
 metro_palettes <- list(
-  `full` = metro_colors(),
-  `ice` = metro_colors("purple", "deep purple", "blue", "light blue"),
-  `rainbow` = metro_colors(
+  full = metro_colors(),
+  ice = metro_colors("purple", "deep purple", "blue", "light blue"),
+  gradient = metro_colors("blue", "orange"),
+  rainbow = metro_colors(
     "purple",
     "deep purple",
     "blue",
@@ -205,9 +222,9 @@ metro_palettes <- list(
     "deep orange",
     "red"
   ),
-  `contrast` = metro_colors("blue", "green", "amber", "purple", "red"),
-  `light` = material_colors("light blue", "red", "yellow", "light green", "orange"),
-  `complement` = metro_colors(
+  contrast = metro_colors("blue", "green", "amber", "purple", "red"),
+  light = material_colors("light blue", "red", "yellow", "light green", "orange"),
+  complement = metro_colors(
     "blue grey",
     "blue",
     "light blue",